# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Management API",
        "description": "API documentation for the Task Management project. Authenticate via /api/auth/login and pass the returned token as `Authorization: Bearer <token>` on protected endpoints.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/reports/projects/{project}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Project report — task counts, completion %, overdue, per-task breakdown",
                "operationId": "reportsProject",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project report",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "project_id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "project_name": {
                                            "type": "string",
                                            "example": "Website Redesign"
                                        },
                                        "total_tasks": {
                                            "type": "integer",
                                            "example": 24
                                        },
                                        "completed_tasks": {
                                            "type": "integer",
                                            "example": 18
                                        },
                                        "pending_tasks": {
                                            "type": "integer",
                                            "example": 6
                                        },
                                        "completion_percentage": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 75
                                        },
                                        "overdue_tasks": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "tasks": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "status": {
                                                        "type": "string"
                                                    },
                                                    "priority": {
                                                        "type": "string"
                                                    },
                                                    "deadline": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "assigned_to": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "hours_logged": {
                                                        "type": "number",
                                                        "format": "float"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/employees/{employee}": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Employee report — assigned tasks, hours logged, avg completion",
                "description": "Employees may view only their own report; managers only employees with tasks in their projects; admins may view any.",
                "operationId": "reportsEmployee",
                "parameters": [
                    {
                        "name": "employee",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee report",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "employee_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "employee_name": {
                                            "type": "string",
                                            "example": "Jane Doe"
                                        },
                                        "total_assigned_tasks": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "completed_tasks": {
                                            "type": "integer",
                                            "example": 8
                                        },
                                        "pending_tasks": {
                                            "type": "integer",
                                            "example": 4
                                        },
                                        "total_hours_logged": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 42.5
                                        },
                                        "avg_completion_time_days": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 3.4
                                        },
                                        "work_logs_submitted": {
                                            "type": "integer",
                                            "example": 18
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ForbiddenError"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthorizedError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "ForbiddenError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
            "NotFoundError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Not found"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum-Token",
                "scheme": "bearer"
            }
        }
    },
...
}